# [WIP] Adaptive fragment sizes in Parquet writer