--- a/cpp/src/io/parquet/writer_impl.cu
+++ b/cpp/src/io/parquet/writer_impl.cu
@@ -92,9 +92,9 @@ size_type column_size(column_view const& column, rmm::cuda_stream_view stream)
   if (column.num_children() == 0) { return size_of(column.type()) * column.size(); }
 
   if (column.type().id() == type_id::STRING) {
-    auto scol         = strings_column_view(column);
-    size_type colsize = cudf::detail::get_value<size_type>(scol.offsets(), column.size(), stream);
-    return colsize;
+    auto scol = strings_column_view(column);
+    return cudf::detail::get_value<size_type>(scol.offsets(), column.size(), stream) -
+           cudf::detail::get_value<size_type>(scol.offsets(), 0, stream);
   } else if (column.type().id() == type_id::STRUCT) {
     auto scol     = structs_column_view(column);
     size_type ret = 0;