## Description
Trying to write Parquet files where rows are very wide can result in pages that are much too large due to the default fragment size of 5000.  This in turn can have an adverse effect on file size when using Zstandard compression.  This PR attempts to address this by modifying the fragment size to a value where each fragment will still fit in the desired page size.

Fixes #12613

<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->

## Checklist
- [x] I am familiar with the [Contributing Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [ ] New or existing tests cover these changes.
- [ ] The documentation is up to date with these changes.