@etseidl 2023-01-26
@vuule you had mentioned awhile back that you thought there was a function somewhere that calculates a column's size.  I couldn't find it so implemented my own.  If you can point me to something better I'll use it.