@etseidl 2023-01-26
Side benefit is better performance in benchmarks for list columns, without the performance regression when just setting the default fragment size to 1000.

Before this PR
```
## parquet_write_encode

### [0] NVIDIA RTX A6000

| data_type | cardinality | run_length | Samples |  CPU Time  | Noise |  GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|-------------|------------|---------|------------|-------|------------|-------|------------------|-------------------|-------------------|
|  INTEGRAL |           0 |          1 |      5x | 111.840 ms | 0.32% | 111.835 ms | 0.31% |       4800565202 |         2.146 GiB |       498.123 MiB |
|  INTEGRAL |        1000 |          1 |     12x |  45.089 ms | 0.38% |  45.083 ms | 0.38% |      11908481790 |         2.770 GiB |       161.438 MiB |
|  INTEGRAL |           0 |         32 |     21x |  34.929 ms | 0.50% |  34.924 ms | 0.50% |      15372537807 |         2.770 GiB |        27.720 MiB |
|  INTEGRAL |        1000 |         32 |     17x |  29.901 ms | 0.14% |  29.896 ms | 0.14% |      17957667626 |         2.770 GiB |        14.403 MiB |
|     FLOAT |           0 |          1 |      5x | 106.580 ms | 0.08% | 106.575 ms | 0.08% |       5037494184 |         1.100 GiB |       510.303 MiB |
|     FLOAT |        1000 |          1 |    112x |  30.835 ms | 0.58% |  30.829 ms | 0.58% |      17414302741 |         1.765 GiB |       110.206 MiB |
|     FLOAT |           0 |         32 |     21x |  23.975 ms | 0.20% |  23.970 ms | 0.20% |      22397993527 |         1.765 GiB |        23.640 MiB |
|     FLOAT |        1000 |         32 |     26x |  19.871 ms | 0.24% |  19.866 ms | 0.24% |      27024975803 |         1.765 GiB |         9.888 MiB |
|   DECIMAL |           0 |          1 |     10x |  51.763 ms | 0.14% |  51.758 ms | 0.14% |      10372678425 |       811.156 MiB |       221.308 MiB |
|   DECIMAL |        1000 |          1 |     28x |  18.359 ms | 0.31% |  18.354 ms | 0.31% |      29251120183 |         1.145 GiB |        48.997 MiB |
|   DECIMAL |           0 |         32 |     33x |  15.604 ms | 0.31% |  15.598 ms | 0.31% |      34418188596 |         1.145 GiB |        10.298 MiB |
|   DECIMAL |        1000 |         32 |     37x |  13.608 ms | 0.32% |  13.603 ms | 0.31% |      39468368025 |         1.145 GiB |         4.717 MiB |
| TIMESTAMP |           0 |          1 |      5x | 131.251 ms | 0.07% | 131.245 ms | 0.07% |       4090588948 |         1.170 GiB |       458.580 MiB |
| TIMESTAMP |        1000 |          1 |     19x |  26.401 ms | 0.37% |  26.396 ms | 0.37% |      20339412646 |         1.474 GiB |        92.808 MiB |
| TIMESTAMP |           0 |         32 |    272x |  27.217 ms | 0.57% |  27.211 ms | 0.57% |      19729735749 |         1.474 GiB |        20.948 MiB |
| TIMESTAMP |        1000 |         32 |     29x |  17.745 ms | 0.42% |  17.740 ms | 0.41% |      30263290620 |         1.474 GiB |         8.718 MiB |
|  DURATION |           0 |          1 |      7x |  79.987 ms | 0.10% |  79.982 ms | 0.10% |       6712426044 |       957.340 MiB |       355.521 MiB |
|  DURATION |        1000 |          1 |     34x |  25.773 ms | 0.50% |  25.768 ms | 0.50% |      20835191262 |         1.474 GiB |        90.214 MiB |
|  DURATION |           0 |         32 |     24x |  21.270 ms | 0.40% |  21.264 ms | 0.40% |      25247594804 |         1.474 GiB |        17.107 MiB |
|  DURATION |        1000 |         32 |     30x |  16.814 ms | 0.48% |  16.809 ms | 0.48% |      31939875607 |         1.474 GiB |         8.113 MiB |
|    STRING |           0 |          1 |      5x | 130.180 ms | 0.22% | 130.174 ms | 0.22% |       4124244709 |         1.342 GiB |       597.486 MiB |
|    STRING |        1000 |          1 |    568x |  26.363 ms | 1.36% |  26.358 ms | 1.36% |      20368583961 |       677.964 MiB |        46.473 MiB |
|    STRING |           0 |         32 |      5x | 130.524 ms | 0.05% | 130.519 ms | 0.05% |       4113363887 |         1.342 GiB |       597.486 MiB |
|    STRING |        1000 |         32 |     35x |  14.426 ms | 0.48% |  14.421 ms | 0.48% |      37228695840 |       677.964 MiB |         8.504 MiB |
|      LIST |           0 |          1 |      5x | 534.754 ms | 0.04% | 534.747 ms | 0.04% |       1003971377 |         1.602 GiB |       498.003 MiB |
|      LIST |        1000 |          1 |      5x | 355.656 ms | 0.03% | 355.649 ms | 0.03% |       1509550797 |         2.752 GiB |       166.640 MiB |
|      LIST |           0 |         32 |      5x | 270.222 ms | 0.05% | 270.217 ms | 0.05% |       1986816725 |         2.752 GiB |        37.257 MiB |
|      LIST |        1000 |         32 |      5x | 274.311 ms | 0.11% | 274.306 ms | 0.11% |       1957200557 |         2.752 GiB |        24.421 MiB |
|    STRUCT |           0 |          1 |      5x | 136.164 ms | 0.09% | 136.159 ms | 0.09% |       3942984412 |         1.283 GiB |       569.525 MiB |
|    STRUCT |        1000 |          1 |    369x |  40.609 ms | 0.63% |  40.603 ms | 0.63% |      13222497588 |         1.324 GiB |        90.699 MiB |
|    STRUCT |           0 |         32 |      5x | 109.019 ms | 0.48% | 109.013 ms | 0.48% |       4924833685 |         1.473 GiB |       409.317 MiB |
|    STRUCT |        1000 |         32 |     19x |  27.182 ms | 0.32% |  27.176 ms | 0.32% |      19755007841 |         1.324 GiB |        15.400 MiB |
```

With these changes:
```
| data_type | cardinality | run_length | Samples |  CPU Time  | Noise |  GPU Time  | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|-----------|-------------|------------|---------|------------|-------|------------|-------|------------------|-------------------|-------------------|
|  INTEGRAL |           0 |          1 |      5x | 111.332 ms | 0.10% | 111.327 ms | 0.10% |       4822460334 |         2.146 GiB |       498.123 MiB |
|  INTEGRAL |        1000 |          1 |     43x |  44.586 ms | 0.50% |  44.581 ms | 0.50% |      12042558020 |         2.770 GiB |       161.438 MiB |
|  INTEGRAL |           0 |         32 |     15x |  34.555 ms | 0.30% |  34.550 ms | 0.29% |      15538786092 |         2.770 GiB |        27.720 MiB |
|  INTEGRAL |        1000 |         32 |     17x |  29.718 ms | 0.45% |  29.712 ms | 0.46% |      18069054636 |         2.770 GiB |        14.403 MiB |
|     FLOAT |           0 |          1 |      5x | 106.595 ms | 0.14% | 106.590 ms | 0.14% |       5036783215 |         1.100 GiB |       510.303 MiB |
|     FLOAT |        1000 |          1 |     17x |  30.503 ms | 0.36% |  30.498 ms | 0.36% |      17603747240 |         1.765 GiB |       110.206 MiB |
|     FLOAT |           0 |         32 |     21x |  23.825 ms | 0.26% |  23.820 ms | 0.26% |      22538378841 |         1.765 GiB |        23.640 MiB |
|     FLOAT |        1000 |         32 |     26x |  19.719 ms | 0.50% |  19.714 ms | 0.50% |      27233359575 |         1.765 GiB |         9.888 MiB |
|   DECIMAL |           0 |          1 |     10x |  51.820 ms | 0.38% |  51.815 ms | 0.38% |      10361253329 |       811.156 MiB |       221.308 MiB |
|   DECIMAL |        1000 |          1 |     28x |  18.281 ms | 0.49% |  18.276 ms | 0.49% |      29375372176 |         1.145 GiB |        48.997 MiB |
|   DECIMAL |           0 |         32 |     33x |  15.538 ms | 0.33% |  15.532 ms | 0.33% |      34564602643 |         1.145 GiB |        10.298 MiB |
|   DECIMAL |        1000 |         32 |    608x |  13.573 ms | 0.58% |  13.567 ms | 0.58% |      39570530602 |         1.145 GiB |         4.717 MiB |
| TIMESTAMP |           0 |          1 |      5x | 129.445 ms | 0.19% | 129.440 ms | 0.19% |       4147628465 |         1.170 GiB |       458.580 MiB |
| TIMESTAMP |        1000 |          1 |     20x |  26.116 ms | 0.21% |  26.111 ms | 0.21% |      20561446173 |         1.474 GiB |        92.808 MiB |
| TIMESTAMP |           0 |         32 |    528x |  26.924 ms | 0.62% |  26.918 ms | 0.62% |      19944485006 |         1.474 GiB |        20.948 MiB |
| TIMESTAMP |        1000 |         32 |     56x |  17.680 ms | 0.50% |  17.674 ms | 0.50% |      30375588647 |         1.474 GiB |         8.718 MiB |
|  DURATION |           0 |          1 |      7x |  79.813 ms | 0.14% |  79.808 ms | 0.14% |       6727042871 |       957.340 MiB |       355.521 MiB |
|  DURATION |        1000 |          1 |     20x |  25.534 ms | 0.22% |  25.529 ms | 0.22% |      21029918810 |         1.474 GiB |        90.214 MiB |
|  DURATION |           0 |         32 |     24x |  21.126 ms | 0.34% |  21.121 ms | 0.34% |      25418552644 |         1.474 GiB |        17.108 MiB |
|  DURATION |        1000 |         32 |    138x |  16.675 ms | 0.50% |  16.670 ms | 0.50% |      32206674557 |         1.474 GiB |         8.113 MiB |
|    STRING |           0 |          1 |      5x | 129.812 ms | 0.24% | 129.807 ms | 0.24% |       4135924927 |         1.342 GiB |       597.486 MiB |
|    STRING |        1000 |          1 |    560x |  26.538 ms | 0.86% |  26.533 ms | 0.86% |      20234323957 |       677.964 MiB |        46.473 MiB |
|    STRING |           0 |         32 |      5x | 130.361 ms | 0.13% | 130.356 ms | 0.13% |       4118511404 |         1.342 GiB |       597.486 MiB |
|    STRING |        1000 |         32 |     72x |  14.652 ms | 0.50% |  14.647 ms | 0.50% |      36654441681 |       677.964 MiB |         8.504 MiB |
|      LIST |           0 |          1 |      9x | 263.635 ms | 0.48% | 263.630 ms | 0.48% |       2036458592 |         1.602 GiB |       498.638 MiB |
|      LIST |        1000 |          1 |      5x | 141.951 ms | 0.21% | 141.945 ms | 0.21% |       3782243064 |         2.752 GiB |       167.795 MiB |
|      LIST |           0 |         32 |     13x |  94.726 ms | 0.50% |  94.721 ms | 0.50% |       5667941431 |         2.752 GiB |        39.319 MiB |
|      LIST |        1000 |         32 |    163x |  92.155 ms | 0.58% |  92.149 ms | 0.58% |       5826129787 |         2.752 GiB |        25.456 MiB |
|    STRUCT |           0 |          1 |      5x | 136.310 ms | 0.28% | 136.304 ms | 0.28% |       3938768656 |         1.283 GiB |       569.525 MiB |
|    STRUCT |        1000 |          1 |     13x |  40.940 ms | 0.32% |  40.935 ms | 0.32% |      13115251265 |         1.324 GiB |        90.700 MiB |
|    STRUCT |           0 |         32 |      5x | 109.057 ms | 0.35% | 109.052 ms | 0.35% |       4923061272 |         1.473 GiB |       409.317 MiB |
|    STRUCT |        1000 |         32 |    528x |  27.551 ms | 0.59% |  27.546 ms | 0.59% |      19490103598 |         1.324 GiB |        15.400 MiB |
```