@etseidl 2023-01-27
@vuule here are the parquet_write_io_compression benchmarks:

before:
```
|    io    | compression | cardinality | run_length | Samples | CPU Time | Noise | GPU Time | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|----------|-------------|-------------|------------|---------|----------|-------|----------|-------|------------------|-------------------|-------------------|
| FILEPATH |      SNAPPY |           0 |          1 |      5x |  3.466 s | 1.42% |  3.466 s | 1.42% |        154889730 |         1.556 GiB |       493.950 MiB |
| FILEPATH |      SNAPPY |        1000 |          1 |      8x |  2.019 s | 1.27% |  2.019 s | 1.27% |        265967925 |         2.536 GiB |       161.238 MiB |
| FILEPATH |      SNAPPY |           0 |         32 |      5x |  1.597 s | 0.35% |  1.597 s | 0.35% |        336229281 |         2.532 GiB |        49.703 MiB |
| FILEPATH |      SNAPPY |        1000 |         32 |     10x |  1.592 s | 0.55% |  1.592 s | 0.55% |        337129111 |         2.536 GiB |        23.416 MiB |
| FILEPATH |        NONE |           0 |          1 |      9x |  1.752 s | 0.74% |  1.752 s | 0.74% |        306399699 |         1.556 GiB |       501.137 MiB |
| FILEPATH |        NONE |        1000 |          1 |     10x |  1.277 s | 0.49% |  1.277 s | 0.49% |        420275736 |         2.536 GiB |       169.774 MiB |
| FILEPATH |        NONE |           0 |         32 |     11x |  1.373 s | 0.75% |  1.373 s | 0.75% |        390934488 |         2.532 GiB |        56.544 MiB |
| FILEPATH |        NONE |        1000 |         32 |      5x |  1.323 s | 0.48% |  1.323 s | 0.48% |        405657348 |         2.536 GiB |        30.410 MiB |
```
This PR:
```
|    io    | compression | cardinality | run_length | Samples | CPU Time | Noise | GPU Time | Noise | bytes_per_second | peak_memory_usage | encoded_file_size |
|----------|-------------|-------------|------------|---------|----------|-------|----------|-------|------------------|-------------------|-------------------|
| FILEPATH |      SNAPPY |           0 |          1 |      5x |  3.485 s | 1.38% |  3.485 s | 1.38% |        154041364 |         1.559 GiB |       494.432 MiB |
| FILEPATH |      SNAPPY |        1000 |          1 |      5x |  2.008 s | 0.32% |  2.008 s | 0.32% |        267352736 |         2.539 GiB |       161.603 MiB |
| FILEPATH |      SNAPPY |           0 |         32 |      5x |  1.598 s | 0.30% |  1.598 s | 0.30% |        336051961 |         2.535 GiB |        50.053 MiB |
| FILEPATH |      SNAPPY |        1000 |         32 |      5x |  1.591 s | 0.21% |  1.591 s | 0.21% |        337388398 |         2.539 GiB |        23.724 MiB |
| FILEPATH |        NONE |           0 |          1 |      5x |  1.742 s | 0.34% |  1.742 s | 0.34% |        308227480 |         1.559 GiB |       501.113 MiB |
| FILEPATH |        NONE |        1000 |          1 |      5x |  1.275 s | 0.42% |  1.275 s | 0.42% |        421134277 |         2.539 GiB |       169.721 MiB |
| FILEPATH |        NONE |           0 |         32 |      5x |  1.357 s | 0.17% |  1.357 s | 0.17% |        395569655 |         2.535 GiB |        56.571 MiB |
| FILEPATH |        NONE |        1000 |         32 |      5x |  1.307 s | 0.20% |  1.307 s | 0.20% |        410754253 |         2.539 GiB |        30.415 MiB |
```